Squashed SOOO many bugs

+++ editusers.php
@@ -1,12 +1,20 @@
 <?php session_start(); ?>
-<?php require 'functions.php'; ?>
 <?php
+require 'functions.php';
 function checkIfValidUser() {
   return (isset($_SESSION["loggedIn"]) and $_SESSION["loggedIn"] and checkUserPassword($_SESSION["user"], $_SESSION["pass"]) and checkIsAdmin($_SESSION["user"], $_SESSION["pass"]));
+}
+function checkIfDBContainsUser($user) {
+  global $DB, $LoginTableName;
+  $results = formatAndQuery("SELECT user FROM %s WHERE user = %sv;", $LoginTableName, $user );
+  return ($results->num_rows > 0);
 }
 function checkPostVarsSet($postData, $expectedKeys) {
   foreach($expectedKeys as $key) {
     if(!isset($postData[$key])) {
+      return False;
+    }
+    if(empty($postData[$key])) {
       return False;
     }
   }
@@ -15,10 +23,10 @@ function checkPostVarsSet($postData, $expectedKeys) {
 if($_SERVER["REQUEST_METHOD"] == "POST" and checkIfValidUser()) {
   $DB = createDBObject();
   if(isset($_POST["formtype"])) {
-    if($_POST["formtype"] == "adduser" and checkPostVarsSet($_POST, array("user", "pass","name"))) {
-      $admin = (isset($_POST["admin"] and $_POST["admin"] == "true")) ? 'TRUE' : 'FALSE';
-      formatAndQuery("INSERT INTO %s VALUES ( %sv, %sv, %sv, %s );",$LoginTableName,$_POST["user"], password_hash($_POST["pass"]), $_POST["name"], $admin);
-    } elseif ($_POST["formtype"] == "deluser" and isset($_POST["user"])) {
+    if($_POST["formtype"] == "adduser" and checkPostVarsSet($_POST, array("user", "pass","name")) and !checkIfDBContainsUser($_POST["user"])) {
+      $admin = (isset($_POST["admin"]) and $_POST["admin"] == "true") ? 'TRUE' : 'FALSE';
+      formatAndQuery("INSERT INTO %s VALUES ( %sv, %sv, %sv, %s );",$LoginTableName,$_POST["user"], password_hash($_POST["pass"], PASSWORD_DEFAULT), $_POST["name"], $admin);
+    } elseif ($_POST["formtype"] == "deluser" and isset($_POST["user"]) and checkIfDBContainsUser($_POST["user"])) {
       formatAndQuery("DELETE FROM %s WHERE user = %sv;", $LoginTableName, $_POST["user"]);
     }
   }
@@ -27,22 +35,25 @@ if($_SERVER["REQUEST_METHOD"] == "POST" and checkIfValidUser()) {
 <html>
 <head>
   <script src="/scripts/jquery-3.1.1.min.js"></script>
+  <link rel="stylesheet" href="<?php global $appdir; echo($appdir);?>styles/edituser.css">
   <script>
-  $(".passbox").click(function checkBoxes(e) {
-  	if ($('#pass1').val() != $('#pass2').val()) {
-  		$('#submit').prop("disabled", true);
-  		$('#errorWarning').show();
-  	} else {
-  		$('#submit').prop("disabled", false);
-  		$('#errorWarning').hide();
-  	}
-  });
-  $(".trashbutton").click(function deleteUser(e) {
-    $user = e.target.data("user");
-    if(window.confirm("Are you sure you want to delete user " + $user + "?")) {
-      $("#deluser").val($user);
-      $("#deluserf").submit();
-    }
+  $(document).ready(function() {
+    $(".passbox").on("input", function checkBoxes(e) {
+    	if ($('#pass1').val() != $('#pass2').val()) {
+    		$('#submit').prop("disabled", true);
+    		$('#errorWarning').show();
+    	} else {
+    		$('#submit').prop("disabled", false);
+    		$('#errorWarning').hide();
+    	}
+    });
+    $(".trashbutton").click(function deleteUser(e) {
+      $user = $(e.target).data("user");
+      if(window.confirm("Are you sure you want to delete user " + $user + "?")) {
+        $("#deluser").val($user);
+        $("#deluserf").submit();
+      }
+    });
   });
   </script>
 </head>
@@ -61,22 +72,26 @@ if($_SERVER["REQUEST_METHOD"] == "POST" and checkIfValidUser()) {
       $results = formatAndQuery("SELECT user, name FROM %s;", $LoginTableName);
       if($results->num_rows > 0) {
         while($data = $results->fetch_assoc()) {
+          echo('<tr>');
           echo('<td>'.clean($data["user"]).'</td>');
           echo('<td>'.clean($data["name"]).'</td>');
-          echo('<img src="/images/red=trash=512.jpg" width="15" class="trashbutton" data-user="'.clean($data["user"]).'">');
+          echo('<td><img src="/images/red-trash-512.jpg" width="15" class="trashbutton" data-user="'.clean($data["user"]).'"></td>');
+          echo('<tr>');
         }
       }
+    } else {
+      echo('<tr><td>You are not logged in</td></tr>');
     }
 
     ?>
     <tr id="addusr">
-      <form id="addUser" action="<?php htmlspecialchars($_SERVER["PHP_SELF"]) ?>" method="post" autocomplete="off">
+      <form id="addUserf" action="<?php htmlspecialchars($_SERVER["PHP_SELF"]) ?>" method="post" autocomplete="off">
         <td>
-          <input id="adduser" name="user" type="text"><input id="pass1" name="pass" type="password" class="passbox"><input id="pass2" type="password" class="passbox">
+          <input id="adduser" name="user" type="text" placeholder="Username"><input id="pass1" name="pass" type="password" class="passbox" placeholder="Password"><input id="pass2" type="password" class="passbox" placeholder="Confirm Password">
           <p id="errorWarning" hidden>PASSWORDS DON'T MATCH</p>
         </td>
-        <td><input id="name" name="name" type="text"><label><input type="checkbox" name="admin" value="true">Admin</label></td>
-        <td><input id="submit" type="image" src="/images/plus-512.ico" width="15"></td>
+        <td><input id="name" name="name" type="text" placeholder = "Name"><label><input type="checkbox" name="admin" value="true">Admin</label></td>
+        <td><input id="submit" type="image" src="/images/plus-4-xxl.png" width="15"></td>
         <input type="hidden" name="formtype" value="adduser">
       </form>
     </tr>
